+++ CHANGELOG.md
@@ -1,7 +1,36 @@
 # Changelog
 
 ## [UNRELEASED]
+
+- support for headers in streamable MCP servers
+- added search to input history
+- added Langfuse OpenTelemetry instrumentation for Agent and Aider
 - fixed last session reload after restart
+
+## [0.23.0]
+
+- fixed laggy prompt field clearing after submitting message
+- added support for streamable remote MCP servers
+- added direct configuration for LM Studio provider
+- added direct configuration for Groq provider
+- improved escape character sanitization in file edit tool
+- render HTML tags properly in messages
+- basic support for udiff format in diff viewer
+- excluding information about disabled power tools in the system prompt
+- improved context files messages in Agent
+- improved Aider initial startup error handling
+
+## [0.22.0]
+
+- added custom commands availability to all modes
+- added info message when agent hits max output tokens
+- improved handling of unknown finish reason during agent iteration
+- fixed UnicodeEncodeError on Windows inside Aider
+- correctly handling stored answer
+- added retry for other finish reason in Agent run
+- todo list disappearance outside agent mode or without useTodoTools enabled
+- added optional target file parameter to /web command
+- /web commands are now saved into input history
 
 ## [0.21.0]